+++ src/earthkit/hydro/__init__.py
@@ -1,3 +1,3 @@
 from .readers import from_cama_downxy, from_cama_nextxy, from_d8, load_river_network, create_river_network
 from .river_network import RiverNetwork
-from .methods import create_subnetwork, accuflux, upstream, downstream, catchment, subcatchment
+from . import accumulation, catchment, subnetwork
